requirements.txt
streamlit==1.31.0
openai==1.12.0
python-dotenv==1.0.0
networkx==3.2.1
pyvis==0.3.2
scikit-learn==1.3.0
pandas==2.1.0
altair==5.0.1
langflow
langchain
chromadb
